requirements.txt
requests==2.32.3
python-telegram-bot==21.9
python-decouple==3.8